requirements.txt
discord
discord.py
python-telegram-bot
openai